+++ client/examples/examples/bessel.hpp
@@ -3,7 +3,7 @@
 // Created Date: 05/11/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 19/05/2021
+// Last Modified: 03/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -17,12 +17,12 @@
 
 using autd::NUM_TRANS_X, autd::NUM_TRANS_Y, autd::TRANS_SPACING_MM;
 
-inline void bessel_test(autd::Controller& autd) {
-  autd.silent_mode() = true;
+inline void bessel_test(autd::ControllerPtr& autd) {
+  autd->silent_mode() = true;
 
   const auto m = autd::modulation::Sine::create(150);  // 150Hz AM
 
   const autd::Vector3 center(TRANS_SPACING_MM * ((NUM_TRANS_X - 1) / 2.0), TRANS_SPACING_MM * ((NUM_TRANS_Y - 1) / 2.0), 150.0);
   const auto g = autd::gain::BesselBeam::create(center, autd::Vector3::UnitZ(), 13.0 / 180.0 * M_PI);
-  autd.send(g, m).unwrap();
+  autd->send(g, m).unwrap();
 }