+++ client/examples/examples/seq.hpp
@@ -3,7 +3,7 @@
 // Created Date: 14/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 19/05/2021
+// Last Modified: 03/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -17,11 +17,11 @@
 
 using autd::NUM_TRANS_X, autd::NUM_TRANS_Y, autd::TRANS_SPACING_MM;
 
-inline void seq_test(autd::Controller& autd) {
-  autd.silent_mode() = false;
+inline void seq_test(autd::ControllerPtr& autd) {
+  autd->silent_mode() = false;
 
   const auto m = autd::modulation::Static::create();
-  autd.send(m).unwrap();
+  autd->send(m).unwrap();
 
   auto seq = autd::sequence::PointSequence::create();
 
@@ -36,5 +36,5 @@ inline void seq_test(autd::Controller& autd) {
 
   const auto actual_freq = seq->set_frequency(1);
   std::cout << "Actual frequency is " << actual_freq << " Hz\n";
-  autd.send(seq).unwrap();
+  autd->send(seq).unwrap();
 }