+++ client/examples/examples/stm.hpp
@@ -3,7 +3,7 @@
 // Created Date: 05/11/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 22/05/2021
+// Last Modified: 03/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -17,11 +17,13 @@
 
 using autd::NUM_TRANS_X, autd::NUM_TRANS_Y, autd::TRANS_SPACING_MM;
 
-inline void stm_test(autd::Controller& autd) {
-  autd.silent_mode() = true;
+inline void stm_test(autd::ControllerPtr& autd) {
+  autd->silent_mode() = true;
 
   const auto m = autd::modulation::Static::create(255);
-  autd.send(nullptr, m).unwrap();
+  autd->send(nullptr, m).unwrap();
+
+  auto stm = autd->stm();
 
   const autd::Vector3 center(TRANS_SPACING_MM * ((NUM_TRANS_X - 1) / 2.0), TRANS_SPACING_MM * ((NUM_TRANS_Y - 1) / 2.0), 150.0);
   const auto point_num = 100;
@@ -30,8 +32,14 @@ inline void stm_test(autd::Controller& autd) {
     const auto theta = 2.0 * M_PI * static_cast<double>(i) / point_num;
     const autd::Vector3 pos(radius * cos(theta), radius * sin(theta), 0.0);
     const auto g = autd::gain::FocalPoint::create(center + pos);
-    autd.stm()->add_gain(g);
+    stm->add_gain(g).unwrap();
   }
 
-  autd.stm()->start(0.5).unwrap();  // 0.5 Hz
+  stm->start(0.5).unwrap();  // 0.5 Hz
+
+  std::cout << "press any key to stop..." << std::endl;
+  std::cin.ignore();
+
+  stm->stop().unwrap();
+  stm->finish().unwrap();
 }