+++ client/include/controller.hpp
@@ -3,7 +3,7 @@
 // Created Date: 05/11/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 20/05/2021
+// Last Modified: 03/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -28,28 +28,40 @@
 
 namespace autd {
 
+class Controller;
+using ControllerPtr = std::unique_ptr<Controller>;
+
 /**
  * @brief AUTD Controller
  */
 class Controller {
+  class STMTimerCallback;
+
+  struct ControllerProps {
+    friend class Controller;
+    friend class STMController;
+    ControllerProps(const bool silent_mode, const bool reads_fpga_info, const bool seq_mode, const bool force_fan)
+        : _silent_mode(silent_mode), _reads_fpga_info(reads_fpga_info), _seq_mode(seq_mode), _force_fan(force_fan) {}
+    ~ControllerProps() = default;
+    ControllerProps(const ControllerProps& v) noexcept = delete;
+    ControllerProps& operator=(const ControllerProps& obj) = delete;
+    ControllerProps(ControllerProps&& obj) = default;
+    ControllerProps& operator=(ControllerProps&& obj) = default;
+
+   private:
+    [[nodiscard]] uint8_t ctrl_flag() const;
+
+    bool _silent_mode;
+    bool _reads_fpga_info;
+    bool _seq_mode;
+    bool _force_fan;
+  };
+
  public:
   class STMController;
+  class STMTimer;
 
-  Controller() noexcept
-      : _link(nullptr),
-        _geometry(std::make_shared<core::Geometry>()),
-        _silent_mode(true),
-        _read_fpga_info(false),
-        _seq_mode(false),
-        _config(core::Configuration::get_default_configuration()),
-        _tx_buf(nullptr),
-        _rx_buf(nullptr),
-        _stm(nullptr) {}
-  ~Controller() = default;
-  Controller(const Controller& v) noexcept = delete;
-  Controller& operator=(const Controller& obj) = delete;
-  Controller(Controller&& obj) = delete;
-  Controller& operator=(Controller&& obj) = delete;
+  static ControllerPtr create();
 
   /**
    * @brief Verify the device is properly connected
@@ -59,7 +71,7 @@ class Controller {
   /**
    * @brief Geometry of the devices
    */
-  [[nodiscard]] core::GeometryPtr geometry() const noexcept;
+  [[nodiscard]] core::GeometryPtr& geometry() noexcept;
 
   /**
    * @brief Silent mode
@@ -70,6 +82,11 @@ class Controller {
    * @brief If true, the devices return FPGA info in all frames. The FPGA info can be read by fpga_info().
    */
   bool& reads_fpga_info() noexcept;
+
+  /**
+   * @brief If true, the fan will be forced to start.
+   */
+  bool& force_fan() noexcept;
 
   /**
    * @brief FPGA info
@@ -96,7 +113,7 @@ class Controller {
    * @param[in] link Link
    * \return ok(whether succeeded), or err(error message) if unrecoverable error is occurred
    */
-  [[nodiscard]] Error open(const core::LinkPtr& link);
+  [[nodiscard]] Error open(core::LinkPtr link);
 
   /**
    * @brief Synchronize all devices
@@ -126,10 +143,9 @@ class Controller {
   /**
    * @brief Send gain to the device
    * @param[in] gain Gain to display
-   * @param[in] wait_for_sent if true, this function will wait for the data is sent to the devices and processed.
    * \return ok(whether succeeded), or err(error message) if unrecoverable error is occurred
    */
-  [[nodiscard]] Error send(const core::GainPtr& gain, bool wait_for_sent = false);
+  [[nodiscard]] Error send(const core::GainPtr& gain);
 
   /**
    * @brief Send modulation to the device
@@ -142,10 +158,9 @@ class Controller {
    * @brief Send gain and modulation to the device
    * @param[in] gain Gain to display
    * @param[in] mod Amplitude modulation to display
-   * @param[in] wait_for_sent if true, this function will wait for the data is sent to the devices and processed.
    * \return ok(whether succeeded), or err(error message) if unrecoverable error is occurred
    */
-  [[nodiscard]] Error send(const core::GainPtr& gain, const core::ModulationPtr& mod, bool wait_for_sent = false);
+  [[nodiscard]] Error send(const core::GainPtr& gain, const core::ModulationPtr& mod);
 
   /**
    * @brief Send sequence and modulation to the device
@@ -162,78 +177,112 @@ class Controller {
 
   /**
    * \brief return pointer to software spatio-temporal modulation controller.
+   * \details Never use Controller before calling STMController::finish
    */
-  [[nodiscard]] std::shared_ptr<STMController> stm() const;
+  [[nodiscard]] std::unique_ptr<STMController> stm();
 
   /**
    * \brief Software spatio-temporal modulation controller.
    */
   class STMController {
    public:
-    explicit STMController(core::LinkPtr link, core::GeometryPtr geometry, bool* silent_mode, bool* read_fpga_info)
-        : _link(std::move(link)), _geometry(std::move(geometry)), _silent_mode(silent_mode), _read_fpga_info(read_fpga_info) {}
+    friend class Controller;
 
     /**
      * @brief Add gain for STM
      */
-    void add_gain(const core::GainPtr& gain);
-
-    /**
-     * @brief Add gains for STM
-     */
-    void add_gains(const std::vector<core::GainPtr>& gains);
+    [[nodiscard]] Error add_gain(const core::GainPtr& gain) const;
 
     /**
      * @brief Start Spatio-Temporal Modulation
      * @param[in] freq Frequency of STM modulation
      * @details Generate STM modulation by switching gains appended by
-     * add_gain() or add_gains() at the freq. The accuracy depends on the computer, for
+     * add_gain() at the freq. The accuracy depends on the computer, for
      * example, about 1ms on Windows. Note that it is affected by interruptions,
      * and so on.
-     * \return ok(whether succeeded), or err(error message) if unrecoverable error is occurred
+     * \return ok, or err(error message) if unrecoverable error is occurred
      */
     [[nodiscard]] Error start(double freq);
 
     /**
-     * @brief Suspend Spatio-Temporal Modulation
-     * \return ok(whether succeeded), or err(error message) if unrecoverable error is occurred
+     * \brief Suspend Spatio-Temporal Modulation
+     * \return ok, or err(error message) if unrecoverable error is occurred
      */
     [[nodiscard]] Error stop();
 
     /**
      * @brief Finish Spatio-Temporal Modulation
-     * @details Added gains will be removed.
-     * \return ok(whether succeeded), or err(error message) if unrecoverable error is occurred
+     * @details Added gains will be removed. Never use this STMController after calling this function.
+     * \return ok, or err(error message) if unrecoverable error is occurred
      */
-    [[nodiscard]] Error finish();
+    [[nodiscard]] Error finish() const;
 
    private:
-    core::LinkPtr _link;
-    core::GeometryPtr _geometry;
-    std::vector<core::GainPtr> _gains;
-    std::vector<std::unique_ptr<uint8_t[]>> _bodies;
-    std::vector<size_t> _sizes;
-    Timer _timer;
-    std::atomic<bool> _lock;
-    bool* _silent_mode;
-    bool* _read_fpga_info;
+    explicit STMController(Controller* p_cnt, std::unique_ptr<STMTimerCallback> handler)
+        : _p_cnt(p_cnt), _handler(std::move(handler)), _timer(nullptr) {}
+
+    Controller* _p_cnt;
+    std::unique_ptr<STMTimerCallback> _handler;
+    std::unique_ptr<core::Timer<STMTimerCallback>> _timer;
   };
 
  private:
-  [[nodiscard]] Error send_header(core::COMMAND cmd, size_t max_trial = 50) const;
-  [[nodiscard]] Error wait_msg_processed(uint8_t msg_id, size_t max_trial = 200) const;
+  Controller() noexcept
+      : _link(nullptr),
+        _config(core::Configuration::get_default_configuration()),
+        _geometry(std::make_unique<core::Geometry>()),
+        _props(ControllerProps(true, false, false, false)),
+        _tx_buf(nullptr),
+        _rx_buf(nullptr) {}
+
+  class STMTimerCallback final : core::CallbackHandler {
+   public:
+    friend class STMController;
+
+    virtual ~STMTimerCallback() = default;
+    STMTimerCallback(const STMTimerCallback& v) noexcept = delete;
+    STMTimerCallback& operator=(const STMTimerCallback& obj) = delete;
+    STMTimerCallback(STMTimerCallback&& obj) = delete;
+    STMTimerCallback& operator=(STMTimerCallback&& obj) = delete;
+
+    explicit STMTimerCallback(core::LinkPtr link) : _link(std::move(link)), _idx(0), _lock(false) {}
+
+    void add(std::unique_ptr<uint8_t[]> data, const size_t size) {
+      this->_bodies.emplace_back(std::move(data));
+      this->_sizes.emplace_back(size);
+    }
+    void clear() {
+      std::vector<std::unique_ptr<uint8_t[]>>().swap(this->_bodies);
+      std::vector<size_t>().swap(this->_sizes);
+      this->_idx = 0;
+    }
+
+    void callback() override {
+      if (auto expected = false; _lock.compare_exchange_weak(expected, true)) {
+        this->_link->send(&this->_bodies[_idx][0], this->_sizes[_idx]).unwrap();
+        this->_idx = (this->_idx + 1) % this->_bodies.size();
+        _lock.store(false, std::memory_order_release);
+      }
+    }
+
+   private:
+    core::LinkPtr _link;
+    std::vector<std::unique_ptr<uint8_t[]>> _bodies;
+    std::vector<size_t> _sizes;
+    size_t _idx;
+    std::atomic<bool> _lock;
+  };
+
+  [[nodiscard]] Error send_header(core::COMMAND cmd) const;
+  [[nodiscard]] Error wait_msg_processed(uint8_t msg_id, size_t max_trial = 50) const;
 
   core::LinkPtr _link;
-  core::GeometryPtr _geometry;
-  bool _silent_mode;
-  bool _read_fpga_info;
-  bool _seq_mode;
   core::Configuration _config;
-
+  core::GeometryPtr _geometry;
+  ControllerProps _props;
   std::unique_ptr<uint8_t[]> _tx_buf;
   std::unique_ptr<uint8_t[]> _rx_buf;
-  std::vector<uint8_t> _fpga_infos;
 
-  std::shared_ptr<STMController> _stm;
+  std::vector<uint8_t> _fpga_infos;
 };
 }  // namespace autd